separate model_cde_utils from script
