+++ .github/workflows/check_new_mdfs.yml
@@ -1,0 +1,42 @@
+name: Check New MDFs
+
+on:
+  schedule:
+    - cron: '0 0 * * *'    # runs daily at midnight UTC (adjust as needed)
+  workflow_dispatch:
+    inputs:
+      models_yaml_path:
+        description: 'Path to YAML files with models & refs their MDFs.'
+        required: false
+        default: "config/mdb_models.yml"
+
+jobs:
+  check_new_mdfs:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v3
+
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.x'
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          version: "0.5.10"
+
+      - name: Install dependencies
+        run: |
+          uv sync --all-extras --dev
+
+      - name: Install project
+        run: |
+          uv pip install .
+
+      - name: Run MDF Check Script
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          python .github/scripts/check_new_mdfs.py
