+++ pyproject.toml
@@ -2,6 +2,7 @@
 name = "bento-mdb-updates"
 version = "0.1.0"
 description = "Tools for updating MDB"
+authors = [{name="Nelson Moore", email="nelson.moore@essential-soft.com"}]
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
