+++ scripts/check_new_mdfs.py
@@ -1,0 +1,2 @@
+"""Check MDF repos for MDFs not in MDB. If found, add to models yaml."""
+
